requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
distro==1.8.0
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
joblib==1.3.2
multidict==6.0.4
nltk==3.8.1
openai==1.2.3
py4j==0.10.9.7
pydantic==2.4.2
pydantic_core==2.10.1
pyspark==3.5.0
python-dotenv==1.0.0
rake-nltk==1.0.6
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
spark-nlp==5.1.4
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
yarl==1.9.2
